feat: Refactor category page and ProjectsList component
